+++ src/main/java/it/gov/pagopa/payhub/pdnd/anpr/connector/AnprApiClientConfig.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.payhub.pdnd.anpr.connector;
 
-import it.gov.pagopa.payhub.pdnd.config.ApiClientConfig;
-import it.gov.pagopa.payhub.pdnd.config.HttpsClientConfig;
+import it.gov.pagopa.payhub.pdnd.config.rest.ApiClientConfig;
+import it.gov.pagopa.payhub.pdnd.config.rest.HttpsClientConfig;
 import it.gov.pagopa.payhub.pdnd.config.pdnd.PdndServiceIntegratedConfig;
 import lombok.AllArgsConstructor;
 import lombok.Data;
