+++ src/test/java/it/gov/pagopa/payhub/pdnd/utils/HttpUtilsTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.payhub.pdnd.utils;
 
-import it.gov.pagopa.payhub.pdnd.config.HttpClientConfig;
+import it.gov.pagopa.payhub.pdnd.config.rest.HttpClientConfig;
 import org.apache.hc.client5.http.HttpRoute;
 import org.apache.hc.client5.http.classic.HttpClient;
 import org.apache.hc.client5.http.config.ConnectionConfig;
